requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
babel==2.18.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
ecdsa==0.19.2
et_xmlfile==2.0.0
fastapi==0.135.3
Flask==3.1.3
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
openai==2.30.0
openpyxl==3.1.5
packaging==26.0
passlib==1.7.4
pefile==2024.8.26
pillow==12.2.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.24
pywin32-ctypes==0.2.3
reportlab==4.4.10
requests==2.33.1
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tkcalendar==1.6.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
Werkzeug==3.1.8
